+++ source/extensions/omni.isaac.ros2_bridge/nodes/OgnROS2CameraHelper.py
@@ -47,17 +47,28 @@ class OgnROS2CameraHelper:
             stage = omni.usd.get_context().get_stage()
             keys = og.Controller.Keys
             with Usd.EditContext(stage, stage.GetSessionLayer()):
-                viewport_api = get_viewport_from_window_name(db.inputs.viewport)
-                if viewport_api:
-                    db.internal_state.viewport = viewport_api
-                    db.internal_state.viewport_name = db.inputs.viewport
-                if db.internal_state.viewport == None:
-                    carb.log_warn("viewport name {db.inputs.viewport} not found")
-                    db.internal_state.initialized = False
-                    return False
+                if db.inputs.viewport:
+                    db.log_warn(
+                        "viewport input is deprecated, please use renderProductPath and the IsaacGetViewportRenderProduct to get a viewports render product path"
+                    )
+                    viewport_api = get_viewport_from_window_name(db.inputs.viewport)
+                    if viewport_api:
+                        db.internal_state.viewport = viewport_api
+                        db.internal_state.viewport_name = db.inputs.viewport
+                    if db.internal_state.viewport == None:
+                        carb.log_warn("viewport name {db.inputs.viewport} not found")
+                        db.internal_state.initialized = False
+                        return False
 
-                viewport = db.internal_state.viewport
-                render_product_prefix = viewport.get_render_product_path().split("/")[-1] + "_"
+                    viewport = db.internal_state.viewport
+                    render_product_path = render_product_path
+                else:
+                    render_product_path = db.inputs.renderProductPath
+                    if stage.GetPrimAtPath(render_product_path) is None:
+                        carb.log_warn("Render product no created yet, retrying on next call")
+                        db.internal_state.initialized = False
+                        return False
+
                 try:
                     if sensor_type == "rgb":
 
@@ -65,7 +76,7 @@ class OgnROS2CameraHelper:
                         submit_node_template_activation(
                             rv + "ROS2PublishImage",
                             0,
-                            [viewport.get_render_product_path()],
+                            [render_product_path],
                             attributes={
                                 "inputs:frameId": db.inputs.frameId,
                                 "inputs:nodeNamespace": db.inputs.nodeNamespace,
@@ -81,7 +92,7 @@ class OgnROS2CameraHelper:
                         submit_node_template_activation(
                             rv + "ROS2PublishImage",
                             0,
-                            [viewport.get_render_product_path()],
+                            [render_product_path],
                             attributes={
                                 "inputs:frameId": db.inputs.frameId,
                                 "inputs:nodeNamespace": db.inputs.nodeNamespace,
@@ -98,13 +109,13 @@ class OgnROS2CameraHelper:
                         submit_node_template_activation(
                             "IsaacReadCameraInfo",
                             0,
-                            [viewport.get_render_product_path()],
+                            [render_product_path],
                             attributes={"inputs:viewport": db.internal_state.viewport_name},
                         )
                         submit_node_template_activation(
                             rv + "ROS2PublishPointCloud",
                             0,
-                            [viewport.get_render_product_path()],
+                            [render_product_path],
                             attributes={
                                 "inputs:frameId": db.inputs.frameId,
                                 "inputs:nodeNamespace": db.inputs.nodeNamespace,
@@ -119,7 +130,7 @@ class OgnROS2CameraHelper:
                         submit_node_template_activation(
                             "ROS2PublishInstanceSegmentation",
                             0,
-                            [viewport.get_render_product_path()],
+                            [render_product_path],
                             attributes={
                                 "inputs:frameId": db.inputs.frameId,
                                 "inputs:nodeNamespace": db.inputs.nodeNamespace,
@@ -134,7 +145,7 @@ class OgnROS2CameraHelper:
                         submit_node_template_activation(
                             "ROS2PublishSemanticSegmentation",
                             0,
-                            [viewport.get_render_product_path()],
+                            [render_product_path],
                             attributes={
                                 "inputs:frameId": db.inputs.frameId,
                                 "inputs:nodeNamespace": db.inputs.nodeNamespace,
@@ -149,7 +160,7 @@ class OgnROS2CameraHelper:
                         submit_node_template_activation(
                             "ROS2PublishBoundingBox2DTight",
                             0,
-                            [viewport.get_render_product_path()],
+                            [render_product_path],
                             attributes={
                                 "inputs:frameId": db.inputs.frameId,
                                 "inputs:nodeNamespace": db.inputs.nodeNamespace,
@@ -163,7 +174,7 @@ class OgnROS2CameraHelper:
                         submit_node_template_activation(
                             "ROS2PublishBoundingBox2DTight",
                             0,
-                            [viewport.get_render_product_path()],
+                            [render_product_path],
                             attributes={
                                 "inputs:frameId": db.inputs.frameId,
                                 "inputs:nodeNamespace": db.inputs.nodeNamespace,
@@ -177,7 +188,7 @@ class OgnROS2CameraHelper:
                         submit_node_template_activation(
                             "ROS2PublishBoundingBox3D",
                             0,
-                            [viewport.get_render_product_path()],
+                            [render_product_path],
                             attributes={
                                 "inputs:frameId": db.inputs.frameId,
                                 "inputs:nodeNamespace": db.inputs.nodeNamespace,
@@ -190,13 +201,13 @@ class OgnROS2CameraHelper:
                         submit_node_template_activation(
                             "IsaacReadCameraInfo",
                             0,
-                            [viewport.get_render_product_path()],
+                            [render_product_path],
                             attributes={"inputs:viewport": db.internal_state.viewport_name},
                         )
                         submit_node_template_activation(
                             "ROS2PublishCameraInfo",
                             0,
-                            [viewport.get_render_product_path()],
+                            [render_product_path],
                             attributes={
                                 "inputs:frameId": db.inputs.frameId,
                                 "inputs:nodeNamespace": db.inputs.nodeNamespace,
@@ -224,7 +235,7 @@ class OgnROS2CameraHelper:
                             submit_node_template_activation(
                                 type_dict[sensor_type] + "ROS2PublishSemanticLabels",
                                 0,
-                                [viewport.get_render_product_path()],
+                                [render_product_path],
                                 attributes={
                                     "inputs:nodeNamespace": db.inputs.nodeNamespace,
                                     "inputs:queueSize": db.inputs.queueSize,